remove deprecated pytest plugin v1
